[코드정리] 간단한 코드 정리 및 canvasSlice 생성 추후 기능 부과예정
