+++ src/store/index.js
@@ -1,13 +1,13 @@
 import {configureStore} from "@reduxjs/toolkit";
-import paletteSlice from "@/store/palette-sclice";
 import choiceSlice from "@/store/choice-slice";
 import tableSlice from "@/store/table-slice";
 import fontSlice from "@/store/font-slice";
+import canvasSlice from "@/store/canvas-sclice";
 
 const store = configureStore({
     reducer:{
         choice: choiceSlice.reducer, 
-        palette: paletteSlice.reducer,
+        canvas: canvasSlice.reducer,
         table: tableSlice.reducer,
         font: fontSlice.reducer
     }
